Remove configuration for MySQL and MySQLi check
A bit of an improvement and future proofing here with the MySQL and
MySQLi selection.

Instead of asking the user of the database to decide whether they should
use MySQL or MySQLi, the script will now detect whether to use MySQL or
MySQLi based on if the functions of the respective connection are
available. The user can still force MySQL to be run by uncommenting Line
74. At some point, this ability will be removed.

Also did a bit of consolidating on where the query to select the
products resides. Since there is only a chance in the connection type
not necessarily the SQL language between MySQL and MySQLi, there is no
need to define this variable twice. Therefore, the lines were
consolidated a bit.