Update LICENSE
Since the coding is being directly incorporated into ZenCart and using their code, plus the unknown nature of it the code CAN be licensed under GPLv3, this license is being replaced with GPLv2. Going forward, with the version marked v2.0, this coding is being licensed with GPLv2. Users of the 1.x series will still be licensed as GPL v3 as that coding doesn't directly come in contact with the original ZenCart coding.